pyproject.toml
[tool.poetry]
name = "gnuradio-pmt"
version = "0.1.0"
description = "gnuradio's pure python pmt converter"
authors = ["crinitusfeles <crinitusfeles@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^25.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
